first pass at fixing #11
The issue is:
When jQuery tries to add 1 to a number with a comma in it, it resets back to zero.

This fix removes the comma out of the number.

Other option I considered was stripping the commas in jQuery before doing the mathematical equations and adding it back after.